Move function to settings repository backend
